Separate initial-state & load-data into 2 files
