+++ src/components/app/index.js
@@ -5,7 +5,8 @@ import store from '../../store';
 import { resetData } from '../../actions';
 import Wrapper from '../wrapper';
 import formatData from '../../utils/format-data';
-import { getInitialState, loadData } from './load-data';
+import getInitialState from './initial-state';
+import loadData from './load-data';
 import '@quantumblack/kedro-ui/lib/styles/app.css';
 import './app.css';
